pyproject.toml
[project]
name = "gooplum-backend"
version = "0.1.0"
description = "Backend for GoopLum project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==24.1.0",
    "deepagents==0.2.7",
    "e2b-code-interpreter==2.3.0",
    "fastapi==0.121.0",
    "langchain==1.1.0",
    "langchain-google-genai==3.1.0",
    "langchain-mcp-adapters==0.1.12",
    "langchain-openai==1.0.2",
    "langgraph==1.0.2",
    "pydantic==2.12.4",
    "python-dotenv==1.2.1",
    "supabase==2.24.0",
]

[dependency-groups]
dev = [
    "black==25.9.0",
    "ruff==0.14.3",
    "langgraph-cli[inmem]==0.4.7",
]
